Struct alloy_signer_aws::signer::AwsSigner

pub struct AwsSigner {
    kms: Client,
    key_id: String,
    pubkey: VerifyingKey,
    address: Address,
    chain_id: Option<ChainId>,
}

Amazon Web Services Key Management Service (AWS KMS) Ethereum signer.

The AWS Signer passes signing requests to the cloud service. AWS KMS keys are identified by a UUID, the key_id.

Because the public key is unknown, we retrieve it on instantiation of the signer. This means that the new function is async and must be called within some runtime.

Note that this signer only supports asynchronous operations. Calling a non-asynchronous method will always return an error.

Examples

use alloy_signer::Signer;
use alloy_signer_aws::AwsSigner;
use aws_config::BehaviorVersion;

let config = aws_config::load_defaults(BehaviorVersion::latest()).await;
let client = aws_sdk_kms::Client::new(&config);

let key_id = "...".to_string();
let chain_id = Some(1);
let signer = AwsSigner::new(client, key_id, chain_id).await.unwrap();

let message = vec![0, 1, 2, 3];

let sig = signer.sign_message(&message).await.unwrap();
assert_eq!(sig.recover_address_from_msg(message).unwrap(), signer.address());

Fields
kms: Clientkey_id: Stringpubkey: VerifyingKeyaddress: Addresschain_id: Option<ChainId>
Implementations

impl AwsSigner

pub async fn new( kms: Client, key_id: String, chain_id: Option<ChainId>, ) -> Result<Self, AwsSignerError>

Instantiate a new signer from an existing Client and key ID.

Retrieves the public key from AWS and calculates the Ethereum address.

pub async fn get_pubkey_for_key( &self, key_id: String, ) -> Result<VerifyingKey, AwsSignerError>

Fetch the pubkey associated with a key ID.

pub async fn get_pubkey(&self) -> Result<VerifyingKey, AwsSignerError>

Fetch the pubkey associated with this signer’s key ID.

pub async fn sign_digest_with_key( &self, key_id: String, digest: &B256, ) -> Result<Signature, AwsSignerError>

Sign a digest with the key associated with a key ID.

pub async fn sign_digest( &self, digest: &B256, ) -> Result<Signature, AwsSignerError>

Sign a digest with this signer’s key

async fn sign_digest_inner( &self, digest: &B256, ) -> Result<Signature, AwsSignerError>

Sign a digest with this signer’s key and applies EIP-155.